fix(TUP-22042): A blank shell is opened when saving job
https://jira.talendforge.org/browse/TUP-22042
I checked the api of Shell on Linux and Windows (Eclipse 4.10)
On Linux, the open method of Shell will open it's parent shell, seems no
way to avoid it.
On Windows, the open method of Shell won't open parent shell.

I think the only way to avoid empty shell on Linux is changing all the
codes which creating empty shell, I think we can even give a null if no
active shell found.